move date range select into comps
